+++ src/tcp_server.cpp
@@ -42,7 +42,7 @@ tcp_server::tcp_server(const stream_info_impl_p &info, const io_context_p &io, c
 	acceptor_->open(protocol);
 
 	// bind to and listen on a free port
-	int port = bind_and_listen_to_port_in_range(*acceptor_,protocol,10);
+	uint16_t port = bind_and_listen_to_port_in_range(*acceptor_,protocol,10);
 
 	// and assign connection-dependent fields
 	// (note: this may be assigned multiple times by multiple TCPs during setup but does not matter)
